backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sse-starlette==2.1.0
anthropic==0.28.0
tavily-python==0.3.3
langgraph==0.1.19
langchain-anthropic==0.1.21
langchain-core==0.2.15
httpx==0.27.0
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
python-multipart==0.0.9
